Refactor of copilotRemoteAgent
