+++ src/view/sessionLogView.ts
@@ -8,8 +8,9 @@ import type * as messages from '../../webviews/sessionLogView/messages';
 import { Disposable, disposeAll } from '../common/lifecycle';
 import { ITelemetry } from '../common/telemetry';
 import { SessionLinkInfo, SessionPullInfo } from '../common/timelineEvent';
+import { IAPISessionLogs } from '../github/common';
 import { CopilotApi, getCopilotApi } from '../github/copilotApi';
-import { CopilotRemoteAgentManager, IAPISessionLogs } from '../github/copilotRemoteAgent';
+import { CopilotRemoteAgentManager } from '../github/copilotRemoteAgent';
 import { CredentialStore } from '../github/credentials';
 import { PullRequestModel } from '../github/pullRequestModel';
 import { PullRequestOverviewPanel } from '../github/pullRequestOverview';
